excelrefine.py
"""
Excel to Markdown Pre-processor (Directory-based batch version)
with .env support
---------------------------------------------------------------
- .envファイルから環境変数を読み込み（FILES_IN_DIR, FILES_OUT_DIR など）。
- FILES_IN_DIR にある複数の .xlsx ファイルを一括処理して、
  処理結果を FILES_OUT_DIR へ出力するサンプルコード。
"""

import os
import glob
import openpyxl
from openpyxl.styles import PatternFill, Font
from openpyxl.utils import get_column_letter

# dotenvの読み込み
from dotenv import load_dotenv


def unmerge_cells_and_annotate(ws):
    """
    【機能】セル結合の情報を展開し、セル値にメタ情報(MergedRange=...)を追記する。
    """
    merged_ranges = list(ws.merged_cells.ranges)
    for merged_range in merged_ranges:
        str_range = str(merged_range)
        top_left_cell = ws.cell(row=merged_range.min_row, column=merged_range.min_col)
        merged_value = top_left_cell.value
        
        # 結合を解除
        ws.unmerge_cells(str_range)

        # 全セルに同じ値を複製
        for row in range(merged_range.min_row, merged_range.max_row + 1):
            for col in range(merged_range.min_col, merged_range.max_col + 1):
                cell = ws.cell(row=row, column=col)
                cell.value = merged_value if merged_value else ""

        # 左上セルに「キー・バリュー形式」でメタ情報を付加
        # 例: "元の値##MergedRange=A1:C1"
        if merged_value:
            top_left_cell.value = f"{merged_value}##MergedRange={str_range}"
        else:
            top_left_cell.value = f"##MergedRange={str_range}"


def extract_conditional_formatting_info(ws):
    """
    【機能】条件付き書式などの情報を取得し、シート末尾にまとめてテキスト化する。
    """
    if not ws.conditional_formatting:
        return
    last_row = ws.max_row + 2
    ws.cell(row=last_row, column=1, value="---ConditionalFormattingInfo---")
    row_offset = 1

    for rule in ws.conditional_formatting:
        rule_str = str(rule)
        info_str = f"CFRule={rule_str}"
        ws.cell(row=last_row + row_offset, column=1, value=info_str)
        row_offset += 1


def store_cell_style_info(ws):
    """
    【機能】セルごとの背景色、フォント情報をキー・バリュー形式で右隣のセルに記録する。
    例: "BG=FFFF00;FontBold=True;FontColor=FF0000"
    """
    for row in ws.iter_rows():
        for cell in row:
            fill_color = None
            if cell.fill and isinstance(cell.fill, PatternFill):
                fill_color = cell.fill.fgColor.value if cell.fill.fgColor else None

            font_info = []
            if cell.font:
                if cell.font.bold:
                    font_info.append("FontBold=True")
                if cell.font.color and cell.font.color.rgb:
                    font_info.append(f"FontColor={cell.font.color.rgb}")

            style_parts = []
            if fill_color:
                style_parts.append(f"BG={fill_color}")
            if font_info:
                style_parts.extend(font_info)

            if style_parts:
                style_text = ";".join(style_parts)
                style_cell = ws.cell(row=cell.row, column=cell.column + 1)
                if style_cell.value:
                    style_cell.value = f"{style_cell.value};{style_text}"
                else:
                    style_cell.value = style_text


def store_comments(ws):
    """
    【機能】セルコメントをキー・バリュー形式でセル値に追記する。
    例: "元の値##Comment=コメント本文"
    """
    for row in ws.iter_rows():
        for cell in row:
            if cell.comment:
                comment_text = cell.comment.text
                current_value = cell.value if cell.value else ""
                new_value = f"{current_value}##Comment={comment_text}"
                cell.value = new_value
                cell.comment = None


def handle_images_shapes(ws):
    """
    【機能】画像・図形などの情報をシート末尾にまとめて記録する。
    """
    images = getattr(ws, '_images', [])
    if images:
        last_row = ws.max_row + 2
        ws.cell(row=last_row, column=1, value="---EmbeddedImagesInfo---")
        row_offset = 1
        for img in images:
            img_info = f"ImageAnchor={img._ref};Width={img.width};Height={img.height}"
            ws.cell(row=last_row + row_offset, column=1, value=img_info)
            row_offset += 1


def annotate_sheet_structure(wb):
    """
    【機能】ワークブック内のシート構成をまとめるサマリシート "Sheet_Structure_Summary" を作成。
    """
    summary_sheet_name = "Sheet_Structure_Summary"
    if summary_sheet_name in wb.sheetnames:
        summary_ws = wb[summary_sheet_name]
    else:
        summary_ws = wb.create_sheet(title=summary_sheet_name)

    summary_ws.cell(row=1, column=1, value="---WorkbookSheetStructure---")
    row_num = 2
    for idx, sheet_name in enumerate(wb.sheetnames, start=1):
        summary_ws.cell(row=row_num, column=1, value=f"SheetIndex={idx};SheetName={sheet_name}")
        row_num += 1


def process_excel_for_markdown(input_file: str, output_file: str):
    """
    【メイン処理】
    1) Excelファイルを読み込む
    2) シートごとに以下の処理を実行
       - セル結合を解除＆注釈付け (##MergedRange=...)
       - 条件付き書式の情報抽出 (CFRule=...)
       - セル装飾情報の抽出 (BG=...;FontBold=...)
       - セルコメントを値に統合 (##Comment=...)
       - 画像/図形メタ情報を抽出 (---EmbeddedImagesInfo---)
    3) 複数シートの構成を要約シートに書き込む (Sheet_Structure_Summary)
    4) 編集後のExcelを出力
    """
    wb = openpyxl.load_workbook(input_file)
    
    # シート構成サマリ作成
    annotate_sheet_structure(wb)

    for ws_name in wb.sheetnames:
        # サマリシート自身の処理はスキップ
        if ws_name == "Sheet_Structure_Summary":
            continue

        ws = wb[ws_name]

        # 1. セル結合を解除＆注釈付け
        unmerge_cells_and_annotate(ws)

        # # 2. 条件付き書式の情報抽出
        # extract_conditional_formatting_info(ws)

        # # 3. セル装飾情報を抜き出して追記
        # store_cell_style_info(ws)

        # # 4. セルコメントを値に統合
        # store_comments(ws)

        # # 5. 画像・図形メタ情報を抽出
        # handle_images_shapes(ws)
    
    wb.save(output_file)
    print(f"Processed Excel saved to {output_file}")


if __name__ == "__main__":
    import sys

    # .envファイルから環境変数を読み込み
    load_dotenv()

    # 環境変数から入出力ディレクトリを取得（なければデフォルト ./input, ./output を使用）
    input_dir = os.getenv("FILES_IN_DIR", "./data-input")
    output_dir = os.getenv("FILES_OUT_DIR", "./data-output")

    # 出力先ディレクトリが存在しない場合は作成しておく
    os.makedirs(output_dir, exist_ok=True)

    # input_dir 内の .xlsx ファイルをすべて取得
    xlsx_files = glob.glob(os.path.join(input_dir, "*.xlsx"))

    if not xlsx_files:
        print(f"[Warning] No .xlsx files found in {input_dir}")
        sys.exit(0)

    # ファイルごとに処理
    for file_path in xlsx_files:
        base_name = os.path.basename(file_path)  # 例: "sample.xlsx"
        output_file = os.path.join(output_dir, base_name)  # 同じファイル名で出力
        print(f"Processing: {file_path} -> {output_file}")
        process_excel_for_markdown(file_path, output_file)
